Major Change in the app structure!
All the proccesses related to the python code - from loading pyodide and packages to running the python code are now done by a web worker. This means that while the calculations are running, the DOM is not blocked and the app remains responsive.

To achieve this, the file pythonRunner was moved to the public directory. It now contains an async function to compute the result matrix/matrices, and it returns a matrix/an array of matrices accordingly.

In addition, the file pythonWorker.js was added to the public directory. This file contains the code for the worker, including loading pyodide and python packages. The loading is done when a transformation page is mounted, and not at the home page. As a result, the file pyodideLoader is no longer needed, as it was previously used to laod pyodide and packages from the entry point of the app.

The file TransformationPage.jsx has several changes as a result of the new structure. When a matrix is submitted, it sends a message to the worker and when the calculation is bak it gets a message back from the worker containing the result.

The other files went minor changes because all the states related to loading pyodide and the python script are no longer needed. This is becuase the main thread is now unaware of pyodide or any python code.